+++ fuzz/fuzz_targets/metering.rs
@@ -1,6 +1,8 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
+mod misc;
+use misc::{SinglePassFuzzModule, ignore_compilation_error, ignore_runtime_error, save_wasm_file};
 use std::sync::Arc;
 use wasmer::wasmparser::Operator;
 use wasmer::{Instance, Module, Store, imports};
@@ -16,16 +18,8 @@ fn cost(operator: &Operator) -> u64 {
     }
 }
 
-fuzz_target!(|module: wasm_smith::Module| {
-    let wasm_bytes = module.to_bytes();
-
-    if let Ok(path) = std::env::var("DUMP_TESTCASE") {
-        use std::fs::File;
-        use std::io::Write;
-        let mut file = File::create(path).unwrap();
-        file.write_all(&wasm_bytes).unwrap();
-        return;
-    }
+fuzz_target!(|module: SinglePassFuzzModule| {
+    let wasm_bytes = module.0.to_bytes();
 
     let mut compiler = Cranelift::default();
     compiler.canonicalize_nans(true);
@@ -33,16 +27,26 @@ fuzz_target!(|module: wasm_smith::Module| {
     let metering = Arc::new(Metering::new(10, cost));
     compiler.push_middleware(metering);
     let mut store = Store::new(compiler);
-    let module = Module::new(&store, &wasm_bytes).unwrap();
+
+    let module = Module::new(&store, &wasm_bytes);
+    let module = match module {
+        Ok(m) => m,
+        Err(e) => {
+            if ignore_compilation_error(&e.to_string()) {
+                return;
+            }
+            save_wasm_file(&wasm_bytes);
+            panic!("{}", e);
+        }
+    };
+
     match Instance::new(&mut store, &module, &imports! {}) {
         Ok(_) => {}
         Err(e) => {
-            let error_message = format!("{}", e);
-            if error_message.starts_with("RuntimeError: ")
-                && error_message.contains("out of bounds")
-            {
+            if ignore_runtime_error(&e.to_string()) {
                 return;
             }
+            save_wasm_file(&wasm_bytes);
             panic!("{}", e);
         }
     }
